+++ package.json
@@ -13,10 +13,13 @@
     "release": "auto shipit"
   },
   "keywords": ["jest", "reporter", "github", "check", "annotation"],
+  "dependencies": {
+    "@octokit/rest": "^16.28.7",
+    "create-check": "^0.2.3",
+    "strip-ansi": "^5.2.0"
+  },
   "devDependencies": {
-    "@types/env-ci": "^3.1.0",
     "@types/jest": "^24.0.17",
-    "@types/lru-cache": "^5.1.0",
     "@types/node": "^12.7.2",
     "@typescript-eslint/eslint-plugin": "^2.0.0",
     "@typescript-eslint/parser": "^2.0.0",
@@ -35,34 +38,18 @@
     "prettier": "^1.18.2",
     "typescript": "^3.5.3"
   },
-  "dependencies": {
-    "@jest/reporters": "^24.9.0",
-    "@octokit/app": "^4.1.0",
-    "@octokit/request": "^5.0.2",
-    "@octokit/rest": "^16.28.7",
-    "env-ci": "^4.1.1",
-    "execa": "^2.0.4",
-    "strip-ansi": "^5.2.0"
-  },
+
   "lint-staged": {
-    "*.{js,css,md}": [
-      "prettier --write",
-      "git add"
-    ]
+    "*.{js,css,md}": ["prettier --write", "git add"]
   },
   "prettier": {
     "singleQuote": true
   },
   "jest": {
     "testLocationInResults": true,
-    "reporters": [
-      "./dist/index.js"
-    ]
+    "reporters": ["./dist/index.js"]
   },
   "auto": {
-    "plugins": [
-      "npm",
-      "released"
-    ]
+    "plugins": ["npm", "released"]
   }
 }
